fix(types): migrate from tsd to vitest for type testing
Migrated the type testing setup from tsd to vitest typecheck to align with
the project's testing infrastructure. This addresses the issue where vitest
wasn't properly configured to handle type definition tests.

Changes:
- Added vitest.config.ts with typecheck configuration
- Renamed test file from index.type-test.ts to index.test-d.ts
- Updated test syntax to use vitest's expectTypeOf API
- Configured tsconfig.test.json to be used by vitest typecheck
- Replaced tsd dependency with vitest in package.json

Fixes #1172
