+++ packages/@secretlint/types/package.json
@@ -50,7 +50,7 @@
     "prepublishOnly": "npm run clean && npm run build",
     "prettier": "prettier --write \"**/*.{js,jsx,ts,tsx,css}\"",
     "prepublish": "npm run --if-present build",
-    "test": "tsc -p ./tsconfig.test.json"
+    "test": "vitest typecheck test/index.test-d.ts --run"
   },
   "prettier": {
     "printWidth": 120,
@@ -61,9 +61,9 @@
     "@types/node": "^24.2.0",
     "prettier": "^2.8.8",
     "tsconfig-to-dual-package": "^1.2.0",
-    "tsd": "^0.33.0",
     "tsx": "^4.20.3",
-    "typescript": "^5.8.3"
+    "typescript": "^5.8.3",
+    "vitest": "^3.2.4"
   },
   "engines": {
     "node": ">=20.0.0"
