+++ packages/@secretlint/types/vitest.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+    test: {
+        typecheck: {
+            enabled: true,
+            only: true,
+            include: ["test/**/*.test-d.ts"],
+            tsconfig: "./tsconfig.test.json",
+        },
+    },
+});
